requirements.txt
requests==2.28.2
python-telegram-bot==20.1
websockets==10.4
pandas==1.5.3
numpy==1.24.2
python-dotenv==1.0.0
aiohttp==3.8.4
asyncio==3.4.3
pytest==7.2.2
logging==0.5.1.2
